--- user-service/src/main/java/com/vroomvroom/user/domain/model/UserStatus.java
@@ -1,7 +1,0 @@
-package com.vroomvroom.user.domain.model;
-
-public enum UserStatus {
-	PENDING,     // 승인 대기
-	APPROVED,    // 승인 완료
-	REJECTED     // 가입 거절
-}
