+++ user-service/src/main/java/com/vroomvroom/userservice/UserServiceApplication.java
@@ -1,0 +1,17 @@
+package com.vroomvroom.userservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+
+@EnableJpaAuditing
+@EnableDiscoveryClient
+@SpringBootApplication
+public class UserServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(UserServiceApplication.class, args);
+	}
+}
