+++ user-service/src/main/java/com/vroomvroom/userservice/domain/model/UserStatus.java
@@ -1,0 +1,7 @@
+package com.vroomvroom.userservice.domain.model;
+
+public enum UserStatus {
+	PENDING,     // 승인 대기
+	APPROVED,    // 승인 완료
+	REJECTED     // 가입 거절
+}
