+++ user-service/src/main/java/com/vroomvroom/userservice/presentation/dto/request/LoginReq.java
@@ -1,0 +1,17 @@
+package com.vroomvroom.userservice.presentation.dto.request;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class LoginReq {
+
+	@Email(message = "이메일 형식이 올바르지 않습니다.")
+	private String email;
+
+	@NotBlank(message = "비밀번호는 필수입니다.")
+	private String password;
+}
