+++ user-service/src/main/java/com/vroomvroom/userservice/security/JwtProperties.java
@@ -1,0 +1,18 @@
+package com.vroomvroom.userservice.security;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+import lombok.Getter;
+
+@Getter
+@Component
+public class JwtProperties {
+
+	@Value("${spring.jwt.access-expiration}")
+	private long accessTokenValidity;
+
+	@Value("${spring.jwt.refresh-expiration}")
+	private long refreshTokenValidity;
+
+}
